some changes for ecs
status

run two process

fix

remove
