+++ index.js
@@ -1,0 +1,4 @@
+var server = require('./server');
+var load = require('./load');
+server();
+load();
