+++ package.json
@@ -11,7 +11,6 @@
     "async": "^1.2.1",
     "cors": "^2.4.1",
     "express": "^4.0.0",
-    "forever": "^0.15.1",
     "node-crontab": "0.0.8",
     "optimist": "^0.6.1",
     "osmium": "*",
@@ -21,10 +20,7 @@
     "underscore": "^1.7.0"
   },
   "scripts": {
-    "start": "npm run server && npm run load",
-    "server": "./node_modules/.bin/forever start server/server.js && ./node_modules/.bin/forever list",
-    "load": "./node_modules/.bin/forever start load/load.js --num_file=937 --num_directory=28 --pguser=$ReportDTDatabaseUsername --pgpassword=$ReportDTDatabasePassword --pgdatabase=$ReportDTDatabaseDatabaseName && ./node_modules/.bin/forever list",
-    "stop": "./node_modules/.bin/forever stop server/server.js && ./node_modules/.bin/forever stop load/load.js"
+    "start": "node index.js"
   },
   "repository": {
     "type": "git",
